Separate code into files
Separated code into file as required.
