+++ my_list.rb
@@ -1,0 +1,12 @@
+require_relative 'my_enumerable'
+
+class MyList
+  include MyEnumerable
+  def initialize(*elements)
+    @list = elements
+  end
+
+  def each(&block)
+    @list.each(&block)
+  end
+end
